+++ core/VoltageSensor.cpp
@@ -2,7 +2,7 @@
 #include "Hardware.h"
 #include "VoltageSensor.h"
 
-#define AnalogToVoltMultiplier 0.004887585532747;
+#define AnalogToVoltMultiplier 0.004887585532747f;
 
 VoltageSensor::VoltageSensor(Hardware *hardware, uint8_t pin) {
     this->hardware = hardware;
@@ -18,9 +18,9 @@ uint8_t VoltageSensor::getPin() {
     return pin;
 }
 
-double VoltageSensor::readVoltage() {
+float VoltageSensor::readVoltage() {
     int analogValue = hardware->readAnalogPin(pin);
-    double voltage = analogValue * AnalogToVoltMultiplier;
+    float voltage = analogValue * AnalogToVoltMultiplier;
     return voltage + interferenceAdjustment;
 }
